Add basic functions for TownResource
TownResource has public variables and no implementation details.

Add getter and setter methods for all variables that need them.
Implement method by which the buy and sell price is calculated.
This calculation is the sum of all individual transactions as if the
selling/buying was occurring one by one. The individual in price is
basePrice * 2^(-decay * stock) and the out price is the same but
stock -> stock-1. Renamed buyPrice and sellPrice to outPrice and inPrice
respectively to reduce confusion.