fixed errors in previous commit, made .pro gui friendly
previous commit failed to compile, and was a console project

the decay variable hadn't had its type declared, I assumed its type was float.
Changed access of the pair values from [0],[1] to .first and .second.
changed the getTownResourcesNeeded return type to be a const since that's what it was.
added the "gui" flag to make file. Added widgets flag to makefile, checking first that the version of QT was compatible.