Added map (contains a lot of magic numbers)
There was no map before this

Added a QGraphicsPixMapItem which will go into a QGraphics scene and allow exploration to be shown.
It has a pointer to its stored pixmap for easier updating since its likely to be updating nearly every frame.
the background is a random beige and exploration is just shown as blackness.