Added World and world scene, Main window displays world scene
Main Window was blank

Main Window now displays a graphicsView which holds the worldScene.
The world scene is controlled by world which also controls all the entities in the game.
Main window runs a timer which is connected to tickGame. This then calls processTicks on World which calls the same on all its children.
The Map can now actually be seen and exploration is imitated by clicking on it.