Added in PlayerSceneItem
There was no representation of the player in the World

Added PlayerSceneItem which represents the location of the player.
currently just drawn with a square. To move the player click and drag to choose a destination.
The processTick method moves the player and explores some of the surrounding map.
World stores a pointer to the object for quick access.