Added townSceneItem skeleton and some methods to World
Towns couldn't exist in the world and were cut off from the rest of the game.

Added townSceneItem but not any methods to it, so it's completely untouched.
Added the addItemToWorld and removeItemFromWorld methods to world. This is the only way items should be added or removed.
Added a getter method for towns to World.
Removed some junk comments from the bottom of resource.h