Added circularSceneItem to optimize collision detection.
Previously collision was detected with path collision which was very expensive

Collision is now calculated as distance < 2*radius which is much faster
checking if an item is under a point is also much faster.
All scene items will be circles of the same size. in future we will add hovering mechanics that shows the specific sprite.