Add TownResource needs initialization
TownResources are constructed in the construct step of Town, but their
townResourcesNeeded is not set. Also the method setTownResourcesNeeded
doesn't change the input, but this is not declared.

Add for loop to add the correct townResourcesNeeded value to each
TownResource in the Town constructor. Mark the parameter for
setTownResourcesNeeded as const.