Add Resource data initialization
There are no actual resources in the code.

Create a  Config class that can have resources and their needs added
into the loadResources function (currently hardcoded here, but could
be changed in future). A given resource and a map of the resources
can be accessed using getResource and getResources respectively.
Initialise with semi dummy resources Food and Iron.