Add preliminary functions for Info class
Representation of information is required for more logic to be
implemented.

Create Info class that holds information about a town. Each entity
(town, player, trader) should have a collection of these (say an
unordered_map with the town name as key) and they should be bought
and sold in each town (value calculation tbd). Each Info instance
currently holds a vector of Resources and stock levels, and the tick
the info was gained. None of these functions are implemented yet.