Add methods to informationHolder, remove duplicate methods from Town
InformationHolder is a skeleton class

Add getAllHeldInfo() which returns all the information.
Add getHeldInfoOnTown which returns a shared pointer to the info or the null pointer if it has no info on the town.
Add addInfo which adds info to the map of held info, replacing information if there was already info for the town.
remove getInfos() from town and the associated member variable, put it InformationHolder.