Add DestinationTown to Trader. Snap movement to Town.
Traders have no concept of where they are going.
Movement completely ignores the existence of towns.

Traders keep a pointer to their destination Town.
This is null if they don't have a destination
Movement is snapped to town if the mouse is within the town.
Moeved movement to its own function to clear up processTick in playerSceneItem.