Add ability to pan with left click and drag. Add sidebar with test button.
Can only change view with scrollbars. Only UI element is the map.

Add ability to click and drag the map. This comes at the expense of being able to draw on the map.
Subclass QGraphicsViewer to overwrite the click and drag cursor so that it looks normal.
Add SideWindow to GUI, this contains just a test button with no function.
Put all GUI elements in QGridLayouts to allow for better layout.