Add inventory on Trader
Traders have no mechanism for moving stock from one town to another
for fun and profit. This is a two part problem, keep track of what the
trader is carrying and provide a means to move these resources to and
from other entities.

Implement a solution to the first problem. Add an inventory property
to the Trader class. This involves the creation of several helper
functions: getInventory, setInventory, adjustInventory,
getInventoryResource, setInventoryResource and adjustInventoryResource.