Add getResource method for Town
Accessing details of a known resource from town is clunky because they
are implemented as a vector.

Create a method getResource that takes a Resource pointer and returns
a TownResource pointer or null pointer if the resource is not found.
Under the hood it iterates through the vector until it finds a
TownResource whose resource is the same as the parameter.