Add buy and sell functions to Trader
Traders cannot transmit their goods from themselves to towns and vice
versa. One of the reasons is a lack of bbuy and sell functions.

Create buy and sell functions on trader. The Sell function assumes
that the given amount is valid and makes the trade. The Buy function
is yet to be written, but should assume that the given amount is valid
from the point of view of the town, but validate against the trader's,
max volume and weight (when implemented) and money.