implement Updating SideWindow and pause in towns
SideWindow has widgets added statically that aren't related to the game state.
Game doesn't pause in the town.

Make the SideWindow update all the buywidgets when the player enters a town.
hide all the buy widgets when the player leaves the town.
pause the game when the player enters a town.
unpause the game when the player leaves a town.
display town name at the top of the sidewindow when the player enters a town.