Activate the ability to buy and sell
The buy and sell buttons send their signals off into the void.

Make the Main menu mediate the purchase since i has access to both
the sideWindow and the playerSceneItem.
Overwrite Trader implementation of buy and sell to in PlayerSceneItem to
call Trader::buy/sell then update its info and emit arrived at town. This updates
the sideWindow to the changes. (consider updating the info in Trader).