FIX: Correct Trader howMuchCanIBuy calculation
The howMuchCanIBuy method returns the amount of a resource a town can
buy with a given amount of money, but we want to use it to calculate
how much a trader can buy from the town.

Correct the method to find the inverse amount of the outPrice method,
rather than the inPrice method. This means traders will no longer buy
more than they can afford. It also means we no longer need to check
intermediate values to prevent overflows.