FIX: Avoid segfault when creating Sell Widget
When creating a widget for the ith resource overall, a signal is
registered for the ith sell widget regardless of whether previous
resources have sell widgets associated with them. This means that if
Iron (i=2) is purchased but there is currently only a widget for food
(i=0), a widget is added to the widget vector for Iron, and then a
signal for widgets[2] is registered. Because widgets only contais two
members, this causes a segfault.

When creating a new sell widget, instead add a signal for the final
member of widgets.