Make Config store Resource vector instead of umap
Config storing an unordered map is inconvenient when it comes to using
it as a reference. We want to be able to rely on all lists of resources
being in the same order. Additionally, with the exception of custom
debugging code, we will almost never need to look up specific
resources, so a map is somewhat unnecessary.

Change Config so m_resources is a vector instead of an unordered map,
and make getResources return value change similarly, make the Resource
pointers const. Change method of inserting needs to work with new
container type.