Construct all BuySellWidgets in SideWindow init
There is much complicated logic and several bugs related to creating
BuySellWidgets as they are required.

Move the construction of all BuySellWidgets into the SideWindow's
constructer. This means we do not pass the stock into the constructor.
We must rely on all resource info in towns being in the same order as
Config's resources and being complete.