Decay value inversely proportional to population
large cities consume and produce more resources, which mean they tend
to have more in stock. This means trading with them tends to be
worthless because they value everything at 0. A town that eats 100 food
a tick and has 1000 in stock should assign it the same value as a town
that eats 1000 food a tick and has 10000.

Scale the decay factor with population. This means as population
increases, rate of decay drops such that the ratio remains the same.
This means that population has to be factored into all value
calculations. Although this is currently handled just by passing the
town's population through all the function calls, it may bbe a better
idea to refactor price calculation entirely, especially as the
functions are called more frequently.