Add inventory screen, FIX config
The inventory tab shows just a label that contains a list of player resources and the amounts.
Each time config is called it provides a different set of resources such that the address can no longer be used as the key.

Add inventory screen which shows a list of player resources as well as their mass, volume, and quantity.
This list is updated whenever the player buys or sells something.
config now has a static vector to resources so that they are only created once.