Add getRelativeDifference method to Resource
There is no quick way of checking what the best deal out there is.
All the comparison functions use multiple exponentials and division.

Create a method getRelativeDifference, which finds the price at which
each party buys a resource, and divides price B by price A. It then
takes the logarithm of the value. Thus a result of 0 means the prices
are equal, 1 means that town B values the resource twice aas much as
town A and so on. The calculations have been simplified on paper,
such that the computer calculation boils down to simple multiplication
and subtraction.