Make AITraders move town if no good deal exists
AITraders do nothing if there are no opportunities to sell the goods
from the current city for a profit.

Keep track of the deals and send the AITrader to the one with the
worst profit to be made. This suggests that there is a decent profit
to be made in the reverse direction.