Add skeleton Carrier class
Traders have to consume resources to move. Different modes of transport
should have different speeds and resource consumption rates. Having a
trader be propelled by a carrier or number of carriers should allow this
behaviour.

Create a Carrier class with a speed and a max speed. This carrier
should consume food at a constant rate, as well as increasing with
speed. Consuming food should increase energy levels and faster speed
should decrease vitality. Lower vitality should reduce the effective
maximum speed. Most of this is yet to be implemented. Also the Carrier
class is currently completely detached from the rest of the code.