Optimize painting for towns, make traders cyan
When displaying the towns price bubble the bounding box and object geometry is updated every frame.
The dark blue AITraders are hard to see on the black background.

Make the bounding box and geometry only update when the info on the town changes,
to this end store the last known info as a mutable value for comparison.
Change the colour of the AITraders cyan.