REFACTOR: Remove InformationHolder class
Although both Trader and Town inherit from InformationHolder, only
Trader actually uses any of its methods and properties.

Roll the InformationHolder class into the Trader class, to reduce
the complexity of the inheritence graph. The functions and properties
should actually probably be a separate class, but then Towns and
Traders should have as a property an instance of that class.
Like instead of a Car class inheriting from an EngineHolder class to
handle running the engine, it should contain an instance of the Engine
class.