REFACTOR: Remove MovableSceneItem class
The set of classes that inherit from the Trader class is the same set
that inherits from the MovableSceneItem class.

Remove the MoveableSceneItem class, and roll its methods and properties
into the Trader class. Now that Traders are QGraphicsObjects we can
assign them signals and slots, which will make future changes easier to
implement. If we later on need a MoveableSceneItem that is not a Trader,
we can break out the methods and properties again, and make Trader
inherit from that instead of CircularSceneObject.