FIX: Update InventoryScreen on inventory change
The Player's inventory screen is only updated when they attempt to buy
or sell a resource. However now a player's inventory can be altered in
other ways, specifically consuming food when on the road.

Create a signal to be released when the player's inventory changes that
updates the Player's inventory screen.