FIX: AITraders only buy food if necessary
Now AITraders won't necessarily be able to sell all their wares in a
town, so they may end the selling phase with food left in their
inventory. Buying 10 more food at this point is unnecessary.

Only buy the amount of food required to bring the total up to 10, if
any purchasing is even required.