FIX: If AITrader can't find a good trade, move on
If an AITrader enters a town with all their assets in resources, but
the town is broke, then the trader can't get any money. As a result,
the trader fails to recognise any trading opportunities.

If the bestOtherTown is a nullptr, find a town that has more money
than the current town and go there, or at least go to the last town
for which you have info.