Only calculate maximum amount if not enough gp
In the process of making a trade, traders calculated both the cost of
the trade and the maximum amount that could be purchased by the
purchasing party. If the purchasing party could buy more than the
traded amount, this is an unnecessary calculation. Indeed in some cases
if the town has enough money and the value of the good is sufficiently
low, it could buy an unlimited amount, which crashes the program in a
function who's value won't even be used.

Only calculate howMuchCanIBuy/Sell if there is insufficient gp to make
the original trade.