REFACTOR: Create InformationHolder class
both Traders and Towns will be holding information. All the information
handling is done in the Trader class.

Move the handling of information to the InformationHolder class. Add an
instance of this class to each trader and (for now) remap each of the
trader's information related functions to it's m_info instance. This
will let us have an instance for Towns as well.