REFACTOR: Improve the maintainability of addTrader function
addTrader function is tightly coupled to the number of closest towns which are added to the traders information.
segfaults if there are less than 4 towns in the world

Replace the manual recording of the closest four towns with a sort. Although this is not a time crucial piece of code this is faster anyway.
This also allows selecting an arbitrary number of closest towns.
Trader now behaves correctly when there are fewer towns than desired.