only display town info when hovering over the town
Town info is displayed whilst hovering over the displayed info bubble as well as the town itself.

Make the info bubble disappear unless the mouse is over the actual town.
This requires special handling of the hoverMoveEvent to stop displaying the bubble if the mouse is not contained by the towns circulars scene item.