Create simple fog, blur edges of vision
vision radius is not obvious when moving around explored areas

create simple fog by drawing a grey image with a transparent circle at the player location.
place fog image at z = 1 so that it obscures AITraders.
TODO: bound the view portal so that the edges of the map and fog images are not seen.