Implement buyInfo and sellInfo
buyInfo and sellInfo are just skeleton declarations.

Implement the details of buying and selling info. The salient points
are that you cannot buy/sell info on a town from/to itself and towns
buy information for the value of the information less the value of
their existing information and sell for the full value.