Remove case for when info is null from getValue
getValue had code specifically for when the object was null

Remove the code from Infos getValue function that pertains to specifically when the info object is null.
This is because dereferencing a null pointer and calling its function is Undefined Behavior.
It just so happens that for many compilers the undefined behavior results in running the code within the function however it is also within its rights to segfault (or do anything else).
Its best that if the function is ever called on a nullptr that program crash so that instances where this can occur are removed.