AITraders to factor destination gp into decisions
AITraders don't take the other Town's available gp into account when
calculating profit. This often leads them to a Town they know is broke
to sell their goods.

As profit is the base2 logarithm of the factor by which they expect to
multiply their money, taking the minimum of that and the logarithm of
the other towns money less the logarithm of their money finds the new
expected factor. Note this doesn't take into account the change in
value based on the stock changing as the trade takes place, but is a
good approximation.