Begin writing method getProfit to get Info value
Information needs to be valued better. An ideal method would be to
calculate the difference in profit between using the old and the new
Info.

Write start of Info::getProfit method, up to calculating what
resources will be bought. It's possible a shortcut exists for
calculating the difference in profit - perhaps the cost of purchasing
the difference between the two sets of shopping lists. Also write the
start of Info::getShoppingList to calculate what resources to buy from
this Town to sell to otherInfo Town in order to maximise profit, given
the trader's gp. When called from getProfit, the purchasing Town's Info
gp can be used, because when trying to maximise profit, a trader will
not spend more money than exists in the coffers of the Town to which
they intend to sell.
Calculating the most efficient way to spend the Trader's gp and gather
the target Town's gp will require further thought, and may require an
approximation.